Add Transformation.map.
